requirements.txt
streamlit>=1.32.0
ezdxf>=1.1.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
pyyaml>=6.0
notion-client>=3.0.0
httpx>=0.27.0
google-auth>=2.28.0
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
